requirements.txt
extcolors==0.1.0
PyMySQL==0.9.2
requests==2.11.1
requests-toolbelt==0.7.0
Scrapy==1.5.0
scrapy-rotating-proxies==0.5